bug 559228 - Update Breakpad to r652
UltraBlame original commit: 22568ff8773cd7cf1ff0b5913265bb70c05e960a

+++ toolkit/crashreporter/google-breakpad/src/client/minidump_file_writer.h
@@ -151,7 +151,7 @@ class UntypedMDRVA {
 
   
   inline MDLocationDescriptor location() const {
-    MDLocationDescriptor location = { size_, position_ };
+    MDLocationDescriptor location = { static_cast<int>(size_), position_ };
     return location;
   }
 
@@ -218,7 +218,7 @@ class TypedMDRVA : public UntypedMDRVA {
   
   
   
-  bool AllocateObjectAndArray(unsigned int count, size_t size);
+  bool AllocateObjectAndArray(size_t count, size_t size);
